+++ HISTORY.rst
@@ -9,6 +9,7 @@ Contributors to this version: Sarah-Claude Bourdeau-Goulet (:user:`Sarahclaude`)
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 * New function hatchmap (:pull:`107`).
+* Support for translating figures. Activating a locale through xclim's ``metadata_locales`` option will try to use metadata saved by xclim or xscen in this locale and to translate common terms appearing in the figures. Figanos currently ships with french translations of those terms. (:pull:`109`, :issue:`64`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^